@@ -1,52 +1,67 @@
 {
-    "name": "sharify",
-    "version": "0.0.1",
-    "private": true,
-    "scripts": {
-        "dev": "vite dev --host",
-        "build": "vite build",
-        "preview": "vite preview",
-        "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
-        "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-        "test": "vitest",
-        "lint": "prettier --check . && eslint .",
-        "format": "prettier --write ."
-    },
-    "devDependencies": {
-        "@apollo/client": "^3.10.4",
-        "@spotify/web-api-ts-sdk": "^1.2.0",
-        "@sveltejs/adapter-auto": "^3.2.0",
-        "@sveltejs/kit": "^2.5.8",
-        "@sveltejs/vite-plugin-svelte": "^3.1.0",
-        "@types/eslint": "^8.56.10",
-        "@types/node": "^20.12.12",
-        "@types/websocket": "^1.0.10",
-        "@typescript-eslint/eslint-plugin": "^7.9.0",
-        "@typescript-eslint/parser": "^7.9.0",
-        "@zerodevx/svelte-toast": "^0.9.5",
-        "autoprefixer": "^10.4.19",
-        "eslint": "^8.57.0",
-        "eslint-config-prettier": "^9.1.0",
-        "eslint-plugin-svelte": "^2.39.0",
-        "graphql": "^16.8.1",
-        "lucide-svelte": "^0.378.0",
-        "postcss": "^8.4.38",
-        "prettier": "^3.2.5",
-        "prettier-plugin-svelte": "^3.2.3",
-        "svelte": "^4.2.17",
-        "svelte-check": "^3.7.1",
-        "tailwindcss": "^3.4.3",
-        "tslib": "^2.6.2",
-        "typescript": "^5.4.5",
-        "vite": "^5.2.11",
-        "vitest": "^1.6.0",
-        "websocket": "^1.0.35"
-    },
-    "type": "module",
-    "dependencies": {
-        "bits-ui": "^0.21.7",
-        "clsx": "^2.1.1",
-        "tailwind-merge": "^2.3.0",
-        "tailwind-variants": "^0.2.1"
-    }
+	"name": "sharify",
+	"private": false,
+	"version": "0.0.1",
+	"type": "module",
+	"scripts": {
+		"dev": "vite dev --host",
+		"build": "vite build",
+		"preview": "vite preview",
+		"prepare": "svelte-kit sync || echo ''",
+		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+		"format": "prettier --write .",
+		"lint": "prettier --check . && eslint .",
+		"test:unit": "vitest",
+		"test": "npm run test:unit -- --run"
+	},
+	"devDependencies": {
+		"@auth/sveltekit": "^1.10.0",
+		"@bufbuild/protobuf": "^2.8.0",
+		"@eslint/compat": "^1.3.2",
+		"@eslint/js": "^9.35.0",
+		"@internationalized/date": "^3.9.0",
+		"@lucide/svelte": "^0.515.0",
+		"@spotify/web-api-ts-sdk": "^1.2.0",
+		"@sveltejs/kit": "^2.42.1",
+		"@sveltejs/vite-plugin-svelte": "^6.2.0",
+		"@tailwindcss/vite": "^4.1.13",
+		"@types/websocket": "^1.0.10",
+		"bits-ui": "^2.9.9",
+		"clsx": "^2.1.1",
+		"eslint": "^9.35.0",
+		"eslint-config-prettier": "^10.1.8",
+		"eslint-plugin-svelte": "^3.12.3",
+		"globals": "^16.4.0",
+		"graphql": "^16.11.0",
+		"lucide-svelte": "^0.525.0",
+		"prettier": "^3.6.2",
+		"prettier-plugin-svelte": "^3.4.0",
+		"prettier-plugin-tailwindcss": "^0.6.14",
+		"svelte": "^5.38.10",
+		"svelte-check": "^4.3.1",
+		"typescript": "^5.9.2",
+		"typescript-eslint": "^8.44.0",
+		"vite": "^7.1.5",
+		"vitest": "^3.2.4"
+	},
+	"pnpm": {
+		"onlyBuiltDependencies": [
+			"@tailwindcss/oxide",
+			"bufferutil",
+			"es5-ext",
+			"esbuild",
+			"utf-8-validate"
+		]
+	},
+	"dependencies": {
+		"@sveltejs/adapter-node": "^5.3.2",
+		"svelte-sonner": "^1.0.5",
+		"tailwind-merge": "^3.3.1",
+		"tailwind-variants": "^2.1.0",
+		"tailwindcss": "^4.1.13",
+		"tw-animate-css": "^1.3.8",
+		"uuid": "^11.1.0",
+		"websocket": "^1.0.35"
+	}
 }
