@@ -1,39 +1,44 @@
 <script lang="ts">
-    import Logo from "$/components/logo.svelte";
-    import { Input } from "$/components/ui/input";
-    import CustomButton from "$/components/button.svelte";
-    import { goto } from "$app/navigation";
+	import Logo from "$/components/logo.svelte";
+	import CustomButton from "$/components/button.svelte";
+	import { goto } from "$app/navigation";
 
-    let room_id = "";
-    let room_password = "";
-    let error = "";
+	let room_id = $state("");
+	let room_password = $state("");
+	let error = $state("");
 
-    async function join_party() {
-        if (room_id.trim() == "" || room_password.trim() == "") {
-            error = "Party ID and Party password must not be empty";
-            return;
-        }
+	async function join_room() {
+		if (room_id.trim().length === 0 || room_password.trim().length === 0) {
+			error = "Room ID and Room password must not be empty";
+			return;
+		}
 
-        await goto(`/join/${room_id}/${room_password}`);
-    }
+		await goto(`/join/${room_id}/${room_password}`);
+	}
 </script>
 
+<svelte:head>
+	<title>Sharify</title>
+</svelte:head>
+
 <section>
-    <Logo class_extension="mb-4" />
-    <Input type="text" placeholder="Party ID" bind:value={room_id} />
-    <Input type="text" placeholder="Party password" bind:value={room_password} />
-    {#if error != ""}
-        <span class="text-red-500">{error}</span>
-    {/if}
-    <CustomButton on:click={join_party}>Join party</CustomButton>
+	<Logo class_extension="mb-4" />
+	<input class="input" type="text" placeholder="Room ID" bind:value={room_id} />
+	<input class="input" type="text" placeholder="Room password" bind:value={room_password} />
+	{#if error !== ""}
+		<span class="text-red-500">{error}</span>
+	{/if}
+	<CustomButton onclick={join_room}>Join room</CustomButton>
 </section>
 
 <style lang="postcss">
-    section {
-        @apply m-auto w-4/12 h-screen flex flex-col gap-6 justify-center items-center;
+	@reference "$/app.css";
 
-        :global(> input) {
-            @apply font-content text-base placeholder:text-main-content text-main-content bg-main-color-hover ring-main-color border-none outline-none w-[25rem];
-        }
-    }
+	section {
+		@apply m-auto flex h-[calc(100vh-var(--nav-h))] w-4/12 flex-col items-center justify-center gap-6;
+
+		input {
+			@apply w-[25rem] border-none bg-main-hover font-content text-base text-main-content ring-main outline-none placeholder:text-main-content;
+		}
+	}
 </style>
