@@ -1,493 +1,282 @@
 import { SpotifyApi } from "@spotify/web-api-ts-sdk";
-import type { PlayHistory, PlaybackState, UserProfile, Queue, SearchResults, Device } from "@spotify/web-api-ts-sdk";
+import type { UserProfile } from "@spotify/web-api-ts-sdk";
 import { env } from "$env/dynamic/public";
 import { writable } from "svelte/store";
 
-import { GetStorageValue, SetStorageValue } from "$/lib/utils";
+import { get_storage_value, set_storage_value, type SpotifyTokens } from "$lib/utils";
 
-type MaxInt<T extends number> = number extends T ? number : _Range<T, []>;
-type _Range<T extends number, R extends unknown[]> = R["length"] extends T
-    ? R[number] | T
-    : _Range<T, [R["length"], ...R]>;
-
-type SpotifyTokens = {
-    access_token: string;
-    refresh_token: string;
-    expires_in: number;
-    created_at: number;
-};
-
-const tokens_default: SpotifyTokens = {
-    access_token: "",
-    refresh_token: "",
-    expires_in: 0,
-    created_at: 0,
+const TOKENS_DEFAULT: SpotifyTokens = {
+	access_token: "",
+	refresh_token: "",
+	expires_in: 0,
+	created_at: 0,
 } as const;
 
-//                     ________  ________  ________  _________  ___  ________ ___    ___ ___  ___  ________  ________   ________  ___       _______   ________
-//                    |\   ____\|\   __  \|\   __  \|\___   ___\\  \|\  _____\\  \  /  /|\  \|\  \|\   __  \|\   ___  \|\   ___ \|\  \     |\  ___ \ |\   __  \
-//                    \ \  \___|\ \  \|\  \ \  \|\  \|___ \  \_\ \  \ \  \__/\ \  \/  / | \  \\\  \ \  \|\  \ \  \\ \  \ \  \_|\ \ \  \    \ \   __/|\ \  \|\  \
-//                     \ \_____  \ \   ____\ \  \\\  \   \ \  \ \ \  \ \   __\\ \    / / \ \   __  \ \   __  \ \  \\ \  \ \  \ \\ \ \  \    \ \  \_|/_\ \   _  _\
-//                      \|____|\  \ \  \___|\ \  \\\  \   \ \  \ \ \  \ \  \_| \/  /  /   \ \  \ \  \ \  \ \  \ \  \\ \  \ \  \_\\ \ \  \____\ \  \_|\ \ \  \\  \|
-//                        ____\_\  \ \__\    \ \_______\   \ \__\ \ \__\ \__\__/  / /      \ \__\ \__\ \__\ \__\ \__\\ \__\ \_______\ \_______\ \_______\ \__\\ _\
-//                       |\_________\|__|     \|_______|    \|__|  \|__|\|__|\___/ /        \|__|\|__|\|__|\|__|\|__| \|__|\|_______|\|_______|\|_______|\|__|\|__|
-//                       \|_________|                                       \|___|/
-
 export class SpotifyHandler {
-    public static readonly BACK_API = `${env.PUBLIC_LOCAL_SERVER_ADDR}/sharify`; // TODO: handle public server addr
-
-    private sdk: SpotifyApi | null = null;
-    private tokens = tokens_default;
-    private client_id = "";
-    private redirectURI = env.PUBLIC_SPOTIFY_REDIRECT;
-    private scopes = [
-        "user-read-private",
-        "user-read-email",
-        "user-modify-playback-state",
-        "user-read-playback-position",
-        "user-library-read",
-        "streaming",
-        "user-read-playback-state",
-        "user-read-recently-played",
-        "playlist-read-private",
-    ];
-    private timeout: ReturnType<typeof setTimeout> | null = null;
-    private code_verifier = "";
-    private code_challenge = "";
-
-    public current_device: Device | null = null;
-    public current_profile: UserProfile | null = null;
-    public is_owner = true;
-    public is_ready = false;
-
-    constructor(client_id: string) {
-        if (client_id == "") {
-            throw new Error("No client id provided");
-        }
-
-        this.client_id = client_id;
-    }
-
-    public InitSdk() {
-        this.sdk = SpotifyApi.withAccessToken(
-            this.client_id,
-            {
-                ...this.tokens,
-                token_type: "code",
-            },
-            { afterRequest: this.HandleRequests },
-        );
-    }
-
-    /**
-     * May throw Error on status 429 rate limiter server side
-     */
-    public async GenerateAuthLink() {
-        try {
-            let res = await fetch(`${SpotifyHandler.BACK_API}/code_verifier`, { method: "GET" });
-            if (res.status == 429) {
-                throw new Error("Error: Too many tries, please try again later");
-            }
-            this.code_verifier = await res.text();
-
-            SetStorageValue({ code_verifier: this.code_verifier });
-
-            res = await fetch(`${SpotifyHandler.BACK_API}/code_challenge/${this.code_verifier}`, { method: "GET" });
-            if (res.status == 429) {
-                throw new Error("Error: Too many tries, please try again later");
-            }
-            this.code_challenge = await res.text();
-
-            const url = new URL("https://accounts.spotify.com/authorize");
-
-            url.search = new URLSearchParams({
-                client_id: this.client_id,
-                response_type: "code",
-                redirect_uri: encodeURI(this.redirectURI),
-                show_dialog: "true",
-                scope: this.scopes.join(" "),
-                code_challenge_method: "S256",
-                code_challenge: this.code_challenge,
-            }).toString();
-
-            return url.toString();
-        } catch (error: any) {
-            console.error(error);
-            return new Error(error);
-        }
-    }
-
-    /**
-     * Can throw Error on failed request, bad status code or empty body
-     */
-    private async CallAuthorizationAPI(body: URLSearchParams) {
-        try {
-            const res = await fetch("https://accounts.spotify.com/api/token", {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/x-www-form-urlencoded",
-                },
-                body,
-            });
-
-            if (res.status != 200) {
-                const error = new Error(
-                    `[CallAuthorizationAPI ${res.status}] Error: ${res.statusText} Response text: ${await res.text()}; Body supplied: ${body}`,
-                );
-                console.error(error);
-                throw error;
-            }
-
-            const json: SpotifyTokens = await res.json();
-
-            if (!json || !json.access_token) {
-                console.error(res);
-                throw new Error(
-                    `[CallAuthorizationAPI] Error: empty body. Response json: ${json}; Body supplied: ${body}`,
-                );
-            }
-
-            this.ProcessTokens(json);
-        } catch (error: any) {
-            console.error(error);
-        }
-    }
-
-    // TODO: Fix refresh token trigger
-    public ProcessTokens(data: Partial<SpotifyTokens>) {
-        if (this.timeout) {
-            clearTimeout(this.timeout);
-        }
-
-        this.tokens = { ...tokens_default, ...data };
-        this.tokens.created_at = data.created_at ?? Date.now();
-
-        // expires_in (usually 3k6s, so) 3600 * 1000 in ms => 1h ; (1000 * 60 * 60) 60 min => 1h as default
-        const expires_in =
-            this.tokens.expires_in && this.tokens.expires_in > 0 ? this.tokens.expires_in * 1000 : 1000 * 60 * 60;
-
-        const msDiff = this.tokens.created_at + expires_in - Date.now();
-        if (msDiff <= 0) {
-            return this.RefreshAccessToken();
-        }
-
-        this.timeout = setTimeout(this.RefreshAccessToken, msDiff);
-        this.TokenFetchingEnded();
-    }
-
-    public FetchAccessToken(code: string) {
-        const code_verifier = GetStorageValue("code_verifier");
-
-        if (code_verifier == null) {
-            throw new Error("No code verifier on local storage");
-        }
-
-        const params = new URLSearchParams({
-            client_id: this.client_id,
-            grant_type: "authorization_code",
-            code: code,
-            redirect_uri: encodeURI(this.redirectURI),
-            code_verifier: code_verifier,
-        });
-
-        this.CallAuthorizationAPI(params);
-    }
-
-    private RefreshAccessToken() {
-        const params = new URLSearchParams({
-            client_id: this.client_id,
-            grant_type: "refresh_token",
-            refresh_token: this.tokens.refresh_token,
-        });
-
-        this.CallAuthorizationAPI(params);
-    }
-
-    public async TokenFetchingEnded() {
-        if (this.tokens.access_token != tokens_default.access_token) {
-            SetStorageValue({
-                st: {
-                    at: this.tokens.access_token,
-                    rt: this.tokens.refresh_token,
-                    ein: this.tokens.expires_in,
-                    date: this.tokens.created_at,
-                },
-            });
-        }
-
-        this.InitSdk();
-        this.is_ready = true;
-
-        const current_device = GetStorageValue("SpotifyDevice");
-
-        try {
-            const [{ devices }, user] = await Promise.all([
-                this.sdk!.player.getAvailableDevices(),
-                this.sdk!.currentUser.profile(),
-            ]);
-
-            this.current_device =
-                (current_device
-                    ? devices.find(device => device.id == current_device.id)
-                    : devices.find(device => device.is_active)) ?? null;
-            this.current_profile = user;
-
-            if (this.current_device) {
-                SetStorageValue({ SpotifyDevice: this.current_device });
-            }
-
-            SetStorageValue({ SpotifyProfile: user });
-        } catch (error) {
-            //location.replace(this.GetAuthLink());
-            console.error(error);
-            return;
-        }
-    }
-
-    private LinkToURI(link: string) {
-        return `spotify:track:${link.split("/").pop()}`.split("?").length > 1
-            ? `spotify:track:${link.split("/").pop()?.split("?").shift()}`
-            : `spotify:track:${link.split("/").pop()}`;
-    }
-
-    public GetTokens() {
-        return this.tokens;
-    }
-
-    private SetTokens() {
-        const tokens = GetStorageValue("st");
-
-        if (tokens == null) {
-            throw new Error("No Spotify tokens on local storage");
-        }
-
-        this.ProcessTokens({
-            access_token: tokens.at,
-            refresh_token: tokens.rt,
-            expires_in: tokens.ein,
-            created_at: tokens.date,
-        });
-    }
-
-    private EnsureInitialized() {
-        if (this.is_ready) {
-            return;
-        }
-
-        this.SetTokens();
-    }
-
-    static ExceededQota<T extends { status: number }>(request: T): boolean {
-        // Handle retry-after HEADER
-
-        return request.status == 429;
-    }
-
-    private HandleRequests(_url: string, _options: RequestInit, response: Response) {
-        if (SpotifyHandler.ExceededQota(response)) {
-            throw new Error("Spotify API Quota excedeed, be patient");
-        }
-    }
-
-    Disconnect() {
-        this.is_ready = false;
-        this.tokens = tokens_default;
-        this.current_device = null;
-        this.current_profile = null;
-        this.sdk = null;
-        this.code_verifier = "";
-        this.code_challenge = "";
-
-        SetStorageValue({ st: null, code_verifier: null, SpotifyProfile: null, SpotifyDevice: null });
-    }
-
-    Pause() {
-        this.EnsureInitialized();
-        return new Promise<SpotifyHandler>(async (resolve, reject) => {
-            try {
-                await this.sdk!.player.pausePlayback(this.current_device?.id ?? "");
-
-                resolve(this);
-            } catch (error: any) {
-                reject(new Error(`There was an error pausing track. (${error})`));
-            }
-        });
-    }
-
-    Resume() {
-        this.EnsureInitialized();
-        return new Promise<SpotifyHandler>(async (resolve, reject) => {
-            try {
-                await this.sdk!.player.startResumePlayback(this.current_device?.id ?? "");
-
-                resolve(this);
-            } catch (error: any) {
-                reject(new Error(`There was an error resuming track. (${error})`));
-            }
-        });
-    }
-
-    AddNextTrack(link: string) {
-        this.EnsureInitialized();
-        return new Promise<void | Error>(async (resolve, reject) => {
-            try {
-                await this.sdk!.player.addItemToPlaybackQueue(this.LinkToURI(link), this.current_device?.id ?? "");
-
-                resolve();
-            } catch (error: any) {
-                reject(new Error(`There was an error adding track to the queue. (${error})`));
-            }
-        });
-    }
-
-    SkipToPrevious() {
-        this.EnsureInitialized();
-        return new Promise<SpotifyHandler>(async (resolve, reject) => {
-            try {
-                await this.sdk!.player.skipToPrevious(this.current_device?.id ?? "");
-
-                resolve(this);
-            } catch (error: any) {
-                reject(new Error(`There was an error skiping to previous track. (${error})`));
-            }
-        });
-    }
-
-    SkipToNext() {
-        this.EnsureInitialized();
-        return new Promise<SpotifyHandler>(async (resolve, reject) => {
-            try {
-                await this.sdk!.player.skipToNext(this.current_device?.id ?? "");
-
-                resolve(this);
-            } catch (error: any) {
-                reject(new Error(`There was an error skiping to next track. (${error})`));
-            }
-        });
-    }
-
-    GetRecentlyPlayedTracks(limit: MaxInt<50>) {
-        this.EnsureInitialized();
-        return new Promise<Array<PlayHistory> | Error>(async (resolve, reject) => {
-            try {
-                const { items } = await this.sdk!.player.getRecentlyPlayedTracks(limit);
-
-                resolve(items);
-            } catch (error: any) {
-                reject(new Error(`There was an error showing the ${limit} recently played tracks. (${error})`));
-            }
-        });
-    }
-
-    SetVolume(value: number) {
-        this.EnsureInitialized();
-        return new Promise<void | Error>(async (resolve, reject) => {
-            try {
-                await this.sdk!.player.setPlaybackVolume(value, this.current_device?.id ?? "");
-
-                resolve();
-            } catch (error: any) {
-                reject(new Error(`There was an error setting volume to ${value}. (${error})`));
-            }
-        });
-    }
-
-    GetCurrentTrackData() {
-        this.EnsureInitialized();
-        return new Promise<PlaybackState | Error>(async (resolve, reject) => {
-            try {
-                const state = await this.sdk!.player.getPlaybackState();
-
-                resolve(state);
-            } catch (error: any) {
-                reject(new Error(`There was an error getting current playback response (${error})`));
-            }
-        });
-    }
-
-    SearchTracks(input: string) {
-        this.EnsureInitialized();
-        return new Promise<SearchResults<Array<"track">> | Error>(async (resolve, reject) => {
-            try {
-                // @ts-ignore
-                const res = await this.sdk!.search(input, ["tracks"], undefined, 10);
-
-                resolve(res);
-            } catch (error: any) {
-                reject(new Error(`There was an error searching for tracks with ${input} (${error})`));
-            }
-        });
-    }
-
-    GetDevices() {
-        this.EnsureInitialized();
-        return new Promise<Array<Device> | Error>(async (resolve, reject) => {
-            try {
-                const { devices } = await this.sdk!.player.getAvailableDevices();
-
-                resolve(devices);
-            } catch (error: any) {
-                reject(new Error(`There was an error searching for devices (${error})`));
-            }
-        });
-    }
-
-    Seek(position: number) {
-        this.EnsureInitialized();
-        return new Promise<void | Error>(async (resolve, reject) => {
-            try {
-                await this.sdk!.player.seekToPosition(position, this.current_device?.id ?? "");
-
-                resolve();
-            } catch (error: any) {
-                reject(new Error(`There was an error seeking to ${position} (${error})`));
-            }
-        });
-    }
-
-    GetProfile(): Promise<UserProfile | Error> {
-        if (!this.is_owner) {
-            this.Disconnect();
-            return new Promise<Error>((_, reject) => reject(new Error()));
-        }
-
-        this.EnsureInitialized();
-        return new Promise<UserProfile | Error>(async (resolve, reject) => {
-            try {
-                const profile = await this.sdk!.currentUser.profile();
-
-                resolve(profile);
-            } catch (error: any) {
-                reject(new Error(`There was an error getting profile (${error})`));
-            }
-        });
-    }
-
-    GetCurrentQueueData() {
-        this.EnsureInitialized();
-        return new Promise<Queue | Error>(async (resolve, reject) => {
-            try {
-                const queue = await this.sdk!.player.getUsersQueue();
-
-                resolve(queue);
-            } catch (error: any) {
-                reject(new Error(`There was an error getting current queue data (${error})`));
-            }
-        });
-    }
-
-    // TODO
-    SetDevice(device: Device) {
-        this.current_device = device;
-    }
+	public static readonly BACK_API = `${env.PUBLIC_SERVER_ADDR_DEV}/v1`; // TODO: handle public server addr
+
+	private sdk: SpotifyApi | null = null;
+	private tokens = TOKENS_DEFAULT;
+	private client_id = "";
+	private redirect_path = "/auth_spotify";
+	private scopes = [
+		"user-read-private",
+		"user-read-email",
+		"user-modify-playback-state",
+		"user-read-playback-position",
+		"user-library-read",
+		"streaming",
+		"user-read-playback-state",
+		"user-read-recently-played",
+		"playlist-read-private",
+	];
+	private timeout: ReturnType<typeof setTimeout> | null = null;
+	private code_verifier = "";
+	private code_challenge = "";
+
+	public current_profile: UserProfile | null = null;
+	public is_ready = false;
+
+	constructor(client_id: string) {
+		if (client_id.length === 0) {
+			throw new Error("No client id provided");
+		}
+
+		this.client_id = client_id;
+	}
+
+	public InitSdk() {
+		this.sdk = SpotifyApi.withAccessToken(
+			this.client_id,
+			{
+				...this.tokens,
+				token_type: "code",
+			},
+			{ afterRequest: this.HandleRequests },
+		);
+	}
+
+	/**
+	 * May throw Error on status 429 rate limiter server side
+	 */
+	public async GenerateAuthLink() {
+		try {
+			let res = await fetch(`${SpotifyHandler.BACK_API}/code_verifier`, { method: "GET" });
+			if (res.status === 429) {
+				throw new Error("Error: Too many tries, please try again later");
+			}
+			this.code_verifier = await res.text();
+
+			set_storage_value({ code_verifier: this.code_verifier });
+
+			res = await fetch(`${SpotifyHandler.BACK_API}/code_challenge/${this.code_verifier}`, { method: "GET" });
+			if (res.status === 429) {
+				throw new Error("Error: Too many tries, please try again later");
+			}
+			this.code_challenge = await res.text();
+
+			const url = new URL("https://accounts.spotify.com/authorize");
+
+			url.search = new URLSearchParams({
+				client_id: this.client_id,
+				response_type: "code",
+				redirect_uri: encodeURI(`${location.origin}${this.redirect_path}`),
+				show_dialog: "true",
+				scope: this.scopes.join(" "),
+				code_challenge_method: "S256",
+				code_challenge: this.code_challenge,
+			}).toString();
+
+			return url.toString();
+		} catch (error) {
+			console.error(error);
+			return new Error(error as string);
+		}
+	}
+
+	private async CallAuthorizationAPI(body: URLSearchParams) {
+		try {
+			const res = await fetch("https://accounts.spotify.com/api/token", {
+				method: "POST",
+				headers: {
+					"Content-Type": "application/x-www-form-urlencoded",
+				},
+				body,
+			});
+
+			if (res.status !== 200) {
+				throw new Error(
+					`[CallAuthorizationAPI ${res.status}] Error: ${res.statusText} Response text: ${await res.text()}; Body supplied: ${body}`,
+				);
+			}
+
+			const json: SpotifyTokens = await res.json();
+
+			if (!json || !json.access_token) {
+				throw new Error(
+					`[CallAuthorizationAPI] Error: empty body. Response json: ${json}; Body supplied: ${body}`,
+				);
+			}
+
+			this.ProcessTokens(json);
+		} catch (error) {
+			console.error(error);
+			this.Disconnect();
+		}
+	}
+
+	public ProcessTokens(data: Partial<SpotifyTokens>) {
+		if (this.timeout) {
+			clearTimeout(this.timeout);
+		}
+
+		this.tokens = { ...TOKENS_DEFAULT, ...data };
+		this.tokens.created_at = data.created_at ?? Date.now();
+
+		// Only handle the Refresh Token at first load, then if it's old, get a new, if its valid,
+		// let the backend handle it
+		//
+		// expires_in (usually 3k6s, so) 3600 * 1000 in ms => 1h ; (1000 * 60 * 60) 60 min => 1h as default
+		const expires_in =
+			this.tokens.expires_in && this.tokens.expires_in > 0 ? this.tokens.expires_in * 1000 : 1000 * 60 * 60;
+
+		const msDiff = this.tokens.created_at + expires_in - Date.now();
+		if (msDiff <= 0) {
+			return this.RefreshAccessToken();
+		}
+
+		this.TokenFetchingEnded();
+	}
+
+	public FetchAccessToken(code: string) {
+		const code_verifier = get_storage_value("code_verifier");
+
+		if (code_verifier === null) {
+			throw new Error("No code verifier on local storage");
+		}
+
+		const params = new URLSearchParams({
+			client_id: this.client_id,
+			grant_type: "authorization_code",
+			code: code,
+			redirect_uri: encodeURI(`${location.origin}${this.redirect_path}`),
+			code_verifier: code_verifier,
+		});
+
+		this.CallAuthorizationAPI(params);
+	}
+
+	private RefreshAccessToken() {
+		const params = new URLSearchParams({
+			client_id: this.client_id,
+			grant_type: "refresh_token",
+			refresh_token: this.tokens.refresh_token,
+		});
+
+		this.CallAuthorizationAPI(params);
+	}
+
+	public async TokenFetchingEnded() {
+		if (this.tokens.access_token !== TOKENS_DEFAULT.access_token) {
+			set_storage_value({ spotify_tokens: this.tokens });
+		}
+
+		this.InitSdk();
+		this.is_ready = true;
+
+		try {
+			const user = await this.sdk!.currentUser.profile();
+
+			this.current_profile = user;
+
+			set_storage_value({ spotify_profile: user });
+		} catch (error) {
+			//location.replace(this.GetAuthLink());
+			console.warn(error);
+
+			this.Disconnect();
+
+			return;
+		}
+	}
+
+	private LinkToURI(link: string) {
+		const id = link.split("/").pop();
+
+		return `spotify:track:${id}`.split("?").length > 1
+			? // Has query
+				`spotify:track:${id?.split("?").shift()}`
+			: `spotify:track:${id}`;
+	}
+
+	public GetTokens() {
+		return this.tokens;
+	}
+
+	public SetTokens() {
+		if (this.is_ready) {
+			return;
+		}
+
+		const tokens = get_storage_value("spotify_tokens");
+
+		if (tokens === null) {
+			console.warn("No Spotify tokens on local storage");
+			return;
+		}
+
+		this.ProcessTokens(tokens);
+	}
+
+	private EnsureInitialized() {
+		if (this.is_ready) {
+			return;
+		}
+
+		this.SetTokens();
+	}
+
+	async GetProfile(): Promise<UserProfile | Error> {
+		this.EnsureInitialized();
+		try {
+			const profile = await this.sdk!.currentUser.profile();
+
+			return profile;
+		} catch (error) {
+			throw new Error(`There was an error getting profile (${error})`);
+		}
+	}
+
+	static ExceededQota<T extends { status: number }>(request: T): boolean {
+		// Handle retry-after HEADER
+
+		return request.status === 429;
+	}
+
+	private HandleRequests(_url: string, _options: RequestInit, response: Response) {
+		if (SpotifyHandler.ExceededQota(response)) {
+			throw new Error("Spotify API Quota excedeed, be patient");
+		}
+	}
+
+	Disconnect() {
+		this.is_ready = false;
+		this.tokens = TOKENS_DEFAULT;
+		this.current_profile = null;
+		this.sdk = null;
+		this.code_verifier = "";
+		this.code_challenge = "";
+
+		set_storage_value({ spotify_tokens: null, code_verifier: null, spotify_profile: null });
+	}
 }
 
 const store = writable<SpotifyHandler | null>(null);
 
-store.update(x => {
-    if (x != null) return x;
+store.update((x) => {
+	if (x !== null) return x;
 
-    return new SpotifyHandler(env.PUBLIC_SPOTIFY_CLIENT_ID!);
+	return new SpotifyHandler(env.PUBLIC_SPOTIFY_CLIENT_ID!);
 });
 
 export default store;
