@@ -1,51 +1,61 @@
 import websocket from "websocket";
-import { type Writable, writable } from "svelte/store";
+import { writable, type Writable } from "svelte/store";
 import { env } from "$env/dynamic/public";
 
-const ws: Writable<websocket.w3cwebsocket | null> = writable(null);
+import { bytes_to_uuid_str } from "./utils";
+import type { Room } from "$/lib/proto/room";
+import type { Nullable } from "./types";
+
+export const ws: Writable<Nullable<websocket.w3cwebsocket>> = writable(null);
 
 function default_onopen() {
-    console.log("WS connected");
+	console.log("WS connected");
 }
 
-function default_onclose() {
-    console.log("WS closed");
+function default_onclose(close_event: websocket.ICloseEvent) {
+	console.log(`WS closed (${JSON.stringify(close_event)})`);
 }
 
 function default_onerror(error: Error) {
-    console.log("WS error", error);
+	console.log("WS error", error);
 }
 
 function default_onmessage(message: websocket.IMessageEvent) {
-    console.log("WS message", message);
+	console.log("WS message", message);
 }
 
 export function init_ws(
-    party_id: number,
-    client_id: number,
-    onopen: typeof default_onopen = default_onopen,
-    onclose: typeof default_onclose = default_onclose,
-    onerror: typeof default_onerror = default_onerror,
-    onmessage: typeof default_onmessage = default_onmessage,
+	room_id: Room["id"],
+	user_id: Room["users"][number]["id"],
+	onopen: typeof default_onopen = default_onopen,
+	onclose: typeof default_onclose = default_onclose,
+	onerror: typeof default_onerror = default_onerror,
+	onmessage: typeof default_onmessage = default_onmessage,
 ) {
-    const _ws = new websocket.w3cwebsocket(
-        `${env.PUBLIC_LOCAL_SERVER_ADDR}/sharify_ws/${party_id}/${client_id}`,
-        undefined,
-        undefined,
-        undefined,
-        undefined,
-        { closeTimeout: 10000 },
-    );
+	const ws_instance = new websocket.w3cwebsocket(
+		`${env.PUBLIC_SERVER_ADDR_DEV}/v1/${bytes_to_uuid_str(room_id)}/${user_id}`, // TODO prod: Change me
+		undefined,
+		undefined,
+		undefined,
+		undefined,
+		{ closeTimeout: 10000 },
+	);
 
-    _ws.onopen = onopen;
+	ws_instance.onopen = () => {
+		ws.update((instance) => {
+			if (instance !== null) {
+				instance.close();
+			}
 
-    _ws.onclose = onclose;
+			return ws_instance;
+		});
 
-    _ws.onerror = onerror;
+		onopen();
+	};
 
-    _ws.onmessage = onmessage;
+	ws_instance.onclose = onclose;
 
-    ws.set(_ws);
+	ws_instance.onerror = onerror;
+
+	ws_instance.onmessage = onmessage;
 }
-
-export default ws;
