@@ -1,7 +1,7 @@
 <script>
-    import { page } from "$app/stores";
+	import { page } from "$app/state";
 </script>
 
 <section>
-    <h1>{$page.status}: {$page.error?.message ?? "Unknown error"}</h1>
+	<h1>{page.status}: {page.error?.message ?? "Unknown error"}</h1>
 </section>
