@@ -1,33 +1,63 @@
 import { redirect } from "@sveltejs/kit";
+import * as uuid from "uuid";
 
+import { PUBLIC_SERVER_ADDR_DEV } from "$env/static/public";
+import { CommandResponse, HttpCommand } from "$lib/proto/cmd";
+import { bytes_to_uuid_str } from "$/lib/utils";
 import type { PageServerLoad } from "./$types";
-import client from "$/lib/server/apollo_client";
-import { GET_PARTY } from "$/lib/queries";
-import type { Party } from "$/lib/types";
 
-export const load: PageServerLoad = async ({ url }) => {
-    const { pathname } = url;
-    const pathname_split = pathname.split("/");
+export const load: PageServerLoad = async ({ url: { pathname } }) => {
+	const pathname_split = pathname.split("/");
 
-    if (pathname_split.length != 4) {
-        return redirect(300, "/");
-    }
+	if (pathname_split.length !== 4) {
+		return redirect(300, "/");
+	}
 
-    const [party_id, password] = pathname_split.slice(2);
+	const [room_id, password] = pathname_split.slice(2);
 
-    const result = await client.query({ query: GET_PARTY, variables: { id: parseInt(party_id) } });
+	if (!uuid.validate(room_id)) {
+		return redirect(301, "/");
+	}
 
-    if (result.error || (result.errors && result.errors.length > 0)) {
-        console.error("Party doesn't exists", result.error, result.errors, party_id);
-        return redirect(301, "/");
-    }
+	const command: HttpCommand = {
+		getRoom: { roomId: uuid.parse(room_id) },
+	};
 
-    const party: Party | null = result.data?.getParty;
+	const bytes = HttpCommand.encode(command).finish();
 
-    if (party == null || party.password != password) {
-        console.error("Party not found or password not right", result.data, party_id, password);
-        return redirect(301, "/");
-    }
+	const res = await fetch(`${PUBLIC_SERVER_ADDR_DEV}/v1`, {
+		method: "POST",
+		headers: {
+			"Content-Type": "application/protobuf",
+		},
+		body: bytes as BodyInit,
+	});
 
-    return { party };
+	if (res.status !== 200 || res.body === null) {
+		console.error(res);
+		return redirect(301, "/");
+	}
+
+	const res_bytes = await res.bytes();
+
+	try {
+		const res_cmd = CommandResponse.decode(res_bytes);
+
+		const room = res_cmd.room;
+
+		if (room === undefined) {
+			console.error(res_cmd);
+			return redirect(301, "/");
+		}
+
+		if (room.password !== password) {
+			console.error("Room password does not match", bytes_to_uuid_str(room.id), password);
+			return redirect(301, "/");
+		}
+
+		return { room };
+	} catch (e: unknown) {
+		console.error(e);
+		return redirect(301, "/");
+	}
 };
