@@ -1,9 +1,26 @@
+import path from "node:path";
+import tailwindcss from "@tailwindcss/vite";
 import { sveltekit } from "@sveltejs/kit/vite";
-import { defineConfig } from "vitest/config";
+import { defineConfig } from "vite";
 
 export default defineConfig({
-    plugins: [sveltekit()],
-    test: {
-        include: ["src/**/*.{test,spec}.{js,ts}"],
-    },
+	plugins: [tailwindcss(), sveltekit()],
+	test: {
+		projects: [
+			{
+				extends: "./vite.config.ts",
+				test: {
+					name: "server",
+					environment: "node",
+					include: ["src/**/*.{test,spec}.{js,ts}"],
+					exclude: ["src/**/*.svelte.{test,spec}.{js,ts}"],
+				},
+			},
+		],
+	},
+	resolve: {
+		alias: {
+			$lib: path.resolve("./src/lib"),
+		},
+	},
 });
